move types to their own folder
