+++ interfaces/index.ts
@@ -1,9 +1,9 @@
 // ANCHOR: Export User Login Type
-export type { userLoginRequestResultInterface } from "./modules/userLoginRequestResultInterface";
-export type { userLoginRequestParametersInterface } from "./modules/userLoginRequestParametersInterface";
-export type { userLoginValidationInterface } from "./modules/userLoginValidationInterface";
+export type { userLoginRequestResultInterface } from "./modules/user/userLoginRequestResultInterface";
+export type { userLoginRequestParametersInterface } from "./modules/user/userLoginRequestParametersInterface";
+export type { userLoginValidationInterface } from "./modules/user/userLoginValidationInterface";
 
 // ANCHOR: Export User Register Type
-export type { userRegisterRequestResultInterface } from "./modules/userRegisterRequestResultInterface";
-export type { userRegisterRequestParametersInterface } from "./modules/userRegisterRequestParametersInterface";
-export type { userRegisterValidationInterface } from "./modules/userRegisterValidationInterface";
+export type { userRegisterRequestResultInterface } from "./modules/user/userRegisterRequestResultInterface";
+export type { userRegisterRequestParametersInterface } from "./modules/user/userRegisterRequestParametersInterface";
+export type { userRegisterValidationInterface } from "./modules/user/userRegisterValidationInterface";
