+++ interfaces/modules/user/userLoginRequestResultInterface.ts
