+++ interfaces/modules/user/userRegisterRequestResultInterface.ts
